+++ include/ibuttons.h
@@ -17,7 +17,7 @@ uint8_t iButton_cmd( uint8_t cmd );
 void iButton_w_bit( int16_t bit );
 uint8_t iButton_reset( void );
 uint8_t check_id( uint8_t *idbuffer);
-uint8_t check_id_2( uint8_t *idbuffer);
+int check_id_2( uint8_t *idbuffer);
 
 
 #endif /* IBUTTONS_H_ */
